POLIO-687: refactor shapes (front-end)
